// src/app/page/master/kategori-aset/kategori-aset.interface.ts
import { EventEmitter } from '@angular/core';

export interface KategoriAset {
    kode?: any;
    kdKategoryAsetHead?: string;
    departemen?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    kdDepartemen?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    id?: any;
    namaDepartemen?: string;
    namaKategoryAset?: string;
   
}